+++ packages/orq-rc/src/orq_ai_sdk/contacts.py
@@ -48,6 +48,8 @@ class Contacts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.CreateContactRequestBody(
             external_id=external_id,
@@ -162,6 +164,8 @@ class Contacts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.CreateContactRequestBody(
             external_id=external_id,