+++ packages/orq-rc/src/orq_ai_sdk/deployments_sdk.py
@@ -54,6 +54,8 @@ class DeploymentsSDK(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.DeploymentsRequest(
             limit=limit,
@@ -160,6 +162,8 @@ class DeploymentsSDK(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.DeploymentsRequest(
             limit=limit,
@@ -305,6 +309,8 @@ class DeploymentsSDK(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.DeploymentGetConfigRequestBody(
             key=key,
@@ -467,6 +473,8 @@ class DeploymentsSDK(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.DeploymentGetConfigRequestBody(
             key=key,
@@ -614,6 +622,8 @@ class DeploymentsSDK(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.Deployments(
             key=key,
@@ -764,6 +774,8 @@ class DeploymentsSDK(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.Deployments(
             key=key,