+++ packages/orq-rc/src/orq_ai_sdk/metrics.py
@@ -64,6 +64,8 @@ class Metrics(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.DeploymentCreateMetricRequest(
             id=id,
@@ -209,6 +211,8 @@ class Metrics(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.DeploymentCreateMetricRequest(
             id=id,