+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -134,7 +134,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-17T13:39:10.688Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-02-19T06:59:00.576Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")