+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -94,5 +94,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-02-17T13:39:11.793Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-02-19T06:59:01.710Z")
     r"""The date and time the resource was created"""