+++ packages/orq-rc/src/orq_ai_sdk/prompts.py
@@ -42,6 +42,8 @@ class Prompts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.GetAllPromptsRequest(
             limit=limit,
@@ -144,6 +146,8 @@ class Prompts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.GetAllPromptsRequest(
             limit=limit,
@@ -244,6 +248,8 @@ class Prompts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         if not isinstance(request, BaseModel):
             request = utils.unmarshal(request, Optional[models.CreatePromptRequestBody])
@@ -345,6 +351,8 @@ class Prompts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         if not isinstance(request, BaseModel):
             request = utils.unmarshal(request, Optional[models.CreatePromptRequestBody])
@@ -444,6 +452,8 @@ class Prompts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.GetOnePromptRequest(
             id=id,
@@ -540,6 +550,8 @@ class Prompts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.GetOnePromptRequest(
             id=id,
@@ -661,6 +673,8 @@ class Prompts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.UpdatePromptRequest(
             id=id,
@@ -811,6 +825,8 @@ class Prompts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.UpdatePromptRequest(
             id=id,
@@ -934,6 +950,8 @@ class Prompts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.DeletePromptRequest(
             id=id,
@@ -1026,6 +1044,8 @@ class Prompts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.DeletePromptRequest(
             id=id,
@@ -1126,6 +1146,8 @@ class Prompts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.ListPromptVersionsRequest(
             prompt_id=prompt_id,
@@ -1231,6 +1253,8 @@ class Prompts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.ListPromptVersionsRequest(
             prompt_id=prompt_id,
@@ -1332,6 +1356,8 @@ class Prompts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.GetPromptVersionRequest(
             prompt_id=prompt_id,
@@ -1437,6 +1463,8 @@ class Prompts(BaseSDK):
 
         if server_url is not None:
             base_url = server_url
+        else:
+            base_url = self._get_url(base_url, url_variables)
 
         request = models.GetPromptVersionRequest(
             prompt_id=prompt_id,