+++ packages/orq-rc/RELEASES.md
@@ -469,3 +469,13 @@ Based on:
 - [python v3.2.0-rc.7] packages/orq-rc
 ### Releases
 - [PyPI v3.2.0-rc.7] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.7 - packages/orq-rc
+
+## 2025-02-19 07:00:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.495.1 (2.515.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0-rc.8] packages/orq-rc
+### Releases
+- [PyPI v3.2.0-rc.8] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.8 - packages/orq-rc